Add files via upload
Insurance data obtained from a private company. Using statistical tools, Docker and R programming language, I used a bayesian approach to tackle the problem of doing inferencial analysis on sinisters numbers and claim amounts. The work was based on simulation methods for computational statistics. The focus was on finding adequate statistical distributions (Poisson and Normal distributions) to explain the data behavior and model it for a future year. 
The impact of this is the robust implementation of a simulated model to calculate accurate individual risk premiums that could be replicated by many competitors in the insurance industry.